Add files via upload
In this string Manipulation program uses of Concatenation which means to join words or sentences.We can use concatenation and Substring is use to extract them. Find is use to find the word on which position it is. By Replace we can replaced the word.By Insert we can insert more words..By Erase we can erase the word..Uses of To Uppercase is to do all alphabets in capital and To Lowercase we can do all alphabets in smaller...